cloud_functions_codes/pda-04-load-bigquery/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-bigquery
google-cloud-logging
psutil
google-cloud-pubsub
pytz